+++ main.cpp
@@ -3,7 +3,7 @@
 #include <string>
 #include <iostream>
 #include "MyMain.h"
-using namespace std;
+
 int main() {
     MyMain M1;
     return 0;